chore(build): raise build targe to the minimum supported, don't bundle utils in snapshots
