+++ packages/browser-playwright/rollup.config.js
@@ -25,7 +25,7 @@ const plugins = [
   json(),
   commonjs(),
   oxc({
-    transform: { target: 'node18' },
+    transform: { target: 'node20' },
   }),
 ]
