+++ packages/browser/rollup.config.js
@@ -37,7 +37,7 @@ const plugins = [
   json(),
   commonjs(),
   oxc({
-    transform: { target: 'node18' },
+    transform: { target: 'node20' },
   }),
 ]
