+++ packages/coverage-v8/rollup.config.js
@@ -31,7 +31,7 @@ const plugins = [
   json(),
   commonjs(),
   oxc({
-    transform: { target: 'node18' },
+    transform: { target: 'node20' },
   }),
 ]
