+++ packages/expect/rollup.config.js
@@ -18,7 +18,7 @@ const dtsUtils = createDtsUtils()
 const plugins = [
   ...dtsUtils.isolatedDecl(),
   oxc({
-    transform: { target: 'node14' },
+    transform: { target: ['node20', 'safari16'] },
   }),
 ]
