+++ packages/runner/rollup.config.js
@@ -25,7 +25,7 @@ const dtsUtils = createDtsUtils()
 const plugins = [
   ...dtsUtils.isolatedDecl(),
   oxc({
-    transform: { target: 'node14' },
+    transform: { target: ['node20', 'safari16'] },
   }),
   json(),
 ]
