+++ packages/snapshot/package.json
@@ -43,12 +43,12 @@
   },
   "dependencies": {
     "@vitest/pretty-format": "workspace:*",
+    "@vitest/utils": "workspace:*",
     "magic-string": "catalog:",
     "pathe": "catalog:"
   },
   "devDependencies": {
     "@types/natural-compare": "^1.4.3",
-    "@vitest/utils": "workspace:*",
     "natural-compare": "^1.4.0"
   }
 }
