+++ packages/snapshot/src/port/state.ts
@@ -17,7 +17,7 @@ import type {
 } from '../types'
 import type { InlineSnapshot } from './inlineSnapshot'
 import type { RawSnapshot, RawSnapshotInfo } from './rawSnapshot'
-import { parseErrorStacktrace } from '../../../utils/src/source-map'
+import { parseErrorStacktrace } from '@vitest/utils/source-map'
 import { saveInlineSnapshots } from './inlineSnapshot'
 import { saveRawSnapshots } from './rawSnapshot'
