+++ packages/snapshot/src/port/utils.ts
@@ -9,8 +9,8 @@ import type { OptionsReceived as PrettyFormatOptions } from '@vitest/pretty-form
 import type { SnapshotData, SnapshotStateOptions } from '../types'
 import type { SnapshotEnvironment } from '../types/environment'
 import { format as prettyFormat } from '@vitest/pretty-format'
+import { isObject } from '@vitest/utils/helpers'
 import naturalCompare from 'natural-compare'
-import { isObject } from '../../../utils/src/helpers'
 import { getSerializers } from './plugins'
 
 // TODO: rewrite and clean up
