+++ packages/ui/rollup.config.js
@@ -47,7 +47,7 @@ export default () => {
         json(),
         commonjs(),
         oxc({
-          transform: { target: 'node18' },
+          transform: { target: 'node20' },
         }),
       ],
       onwarn,
