+++ packages/utils/package.json
@@ -52,10 +52,6 @@
       "types": "./dist/display.d.ts",
       "default": "./dist/display.js"
     },
-    "./highlight": {
-      "types": "./dist/highlight.d.ts",
-      "default": "./dist/highlight.js"
-    },
     "./source-map": {
       "types": "./dist/source-map.d.ts",
       "default": "./dist/source-map.js"
@@ -92,7 +88,6 @@
     "@jridgewell/trace-mapping": "catalog:",
     "@types/estree": "catalog:",
     "diff-sequences": "^29.6.3",
-    "loupe": "^3.2.1",
-    "tinyhighlight": "^0.3.2"
+    "loupe": "^3.2.1"
   }
 }
