+++ packages/vitest/package.json
@@ -218,6 +218,7 @@
     "mime": "^4.1.0",
     "prompts": "^2.4.2",
     "strip-literal": "catalog:",
+    "tinyhighlight": "catalog:",
     "ws": "catalog:"
   }
 }
