+++ packages/ws-client/rollup.config.js
@@ -37,7 +37,7 @@ export default () => [
       json(),
       commonjs(),
       oxc({
-        transform: { target: 'node18' },
+        transform: { target: ['node20', 'safari16'] },
       }),
     ],
   },
