+++ pnpm-workspace.yaml
@@ -80,6 +80,7 @@ catalog:
   std-env: ^3.10.0
   strip-literal: ^3.1.0
   tinyglobby: ^0.2.15
+  tinyhighlight: ^0.3.2
   tinyrainbow: ^3.0.3
   tinyspy: ^4.0.4
   typescript: ^5.9.3
